+++ compiler-plugin/src/main/java/io/ballerina/stdlib/sql/compiler/Utils.java
@@ -31,6 +31,11 @@ import io.ballerina.tools.diagnostics.DiagnosticInfo;
 import java.sql.Types;
 import java.util.Optional;
 
+import static io.ballerina.stdlib.sql.compiler.Constants.TimeRecordTypes.CIVIL;
+import static io.ballerina.stdlib.sql.compiler.Constants.TimeRecordTypes.DATE;
+import static io.ballerina.stdlib.sql.compiler.Constants.TimeRecordTypes.PACKAGE_NAME;
+import static io.ballerina.stdlib.sql.compiler.Constants.TimeRecordTypes.TIME_OF_DAY;
+import static io.ballerina.stdlib.sql.compiler.Constants.TimeRecordTypes.UTC;
 import static io.ballerina.stdlib.sql.compiler.SQLDiagnosticsCodes.SQL_201;
 import static io.ballerina.stdlib.sql.compiler.SQLDiagnosticsCodes.SQL_202;
 import static io.ballerina.stdlib.sql.compiler.SQLDiagnosticsCodes.SQL_203;
@@ -81,7 +86,15 @@ public class Utils {
     }
 
     public static DiagnosticInfo addDiagnosticsForInvalidTypes(String outParameterName,
-                                                               TypeDescKind requestedReturnType) {
+                                                               TypeSymbol argumentTypeSymbol) {
+        TypeDescKind requestedReturnType = argumentTypeSymbol.typeKind();
+        String requestedReturnTypeArgValue = "";
+        if (requestedReturnType == TypeDescKind.TYPE_REFERENCE) {
+            TypeSymbol typeDescriptor = ((TypeReferenceTypeSymbol) argumentTypeSymbol).typeDescriptor();
+            requestedReturnType = typeDescriptor.typeKind();
+            requestedReturnTypeArgValue = argumentTypeSymbol.signature();
+        }
+
         int sqlTypeValue;
         switch (outParameterName) {
             case Constants.OutParameter.VARCHAR:
@@ -232,9 +245,8 @@ public class Utils {
                 }
                 return new DiagnosticInfo(SQL_201.getCode(), SQL_201.getMessage(), SQL_201.getSeverity());
             case Types.DATE:
-                // todo Record validate for time:Date
-                if (requestedReturnType == TypeDescKind.RECORD ||
-                        requestedReturnType == TypeDescKind.TYPE_REFERENCE ||
+                if ((requestedReturnType == TypeDescKind.RECORD && requestedReturnTypeArgValue.startsWith(PACKAGE_NAME)
+                        && requestedReturnTypeArgValue.endsWith(DATE)) ||
                         requestedReturnType == TypeDescKind.INT ||
                         requestedReturnType == TypeDescKind.STRING) {
                     return null;
@@ -242,9 +254,8 @@ public class Utils {
                 return new DiagnosticInfo(SQL_222.getCode(), SQL_222.getMessage(), SQL_222.getSeverity());
             case Types.TIME:
             case Types.TIME_WITH_TIMEZONE:
-                // todo Record validate for time:TimeOfDay
-                if (requestedReturnType == TypeDescKind.RECORD ||
-                        requestedReturnType == TypeDescKind.TYPE_REFERENCE ||
+                if ((requestedReturnType == TypeDescKind.RECORD && requestedReturnTypeArgValue.startsWith(PACKAGE_NAME)
+                        && requestedReturnTypeArgValue.endsWith(TIME_OF_DAY)) ||
                         requestedReturnType == TypeDescKind.INT ||
                         requestedReturnType == TypeDescKind.STRING) {
                     return null;
@@ -252,10 +263,11 @@ public class Utils {
                 return new DiagnosticInfo(SQL_223.getCode(), SQL_223.getMessage(), SQL_223.getSeverity());
             case Types.TIMESTAMP:
             case Types.TIMESTAMP_WITH_TIMEZONE:
-                // todo Record validate for time:Civil, time:Utc
-                if (requestedReturnType == TypeDescKind.RECORD ||
-                        requestedReturnType == TypeDescKind.TYPE_REFERENCE ||
-                        requestedReturnType == TypeDescKind.TUPLE ||
+                if ((requestedReturnType == TypeDescKind.RECORD && requestedReturnTypeArgValue.startsWith(PACKAGE_NAME)
+                        && requestedReturnTypeArgValue.endsWith(CIVIL)) ||
+                    (requestedReturnType == TypeDescKind.INTERSECTION
+                            && requestedReturnTypeArgValue.startsWith(PACKAGE_NAME)
+                            && requestedReturnTypeArgValue.endsWith(UTC)) ||
                         requestedReturnType == TypeDescKind.INT ||
                         requestedReturnType == TypeDescKind.STRING) {
                     return null;