+++ changelog.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [Improve Database columns to Ballerina record Mapping through `sql:Column` Annotation](https://github.com/ballerina-platform/ballerina-standard-library/issues/2652)
 
 ### Changed
+- [Fixed compiler plugin validation for `time` module constructs](https://github.com/ballerina-platform/ballerina-standard-library/issues/2893)
 
 ## [1.3.1] - 2022-03-01